geom viewer: use BrowserModel from RBrowser
Eliminate duplictaed model from geom viewer itself
